+++ ast/continue_stmt.go
@@ -1,0 +1,24 @@
+package ast
+
+import (
+	"github.com/textwire/textwire/token"
+)
+
+type ContinueStmt struct {
+	Token token.Token // The '@continue' token
+}
+
+func (cs *ContinueStmt) statementNode() {
+}
+
+func (cs *ContinueStmt) TokenLiteral() string {
+	return cs.Token.Literal
+}
+
+func (cs *ContinueStmt) String() string {
+	return cs.Token.Literal
+}
+
+func (cs *ContinueStmt) Line() uint {
+	return cs.Token.Line
+}