+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Release notes for 1.x
 
+## v1.2.0 (2024-03-02)
+
+- Added `@break` and `@continue` directives for the `@each` loop.
+- Added `@break` and `@continue` directives for the `@for` loop.
+- Fixed bug where you couldn't compare strings with the `==` operator. For example, `@if(n == "hello")` was returning error.
+
 ## v1.1.0 (2024-02-28)
 
 - Added `@else` directive for the `@each` loop which will be executed if the loop has no items to iterate over.
 - Added `@else` directive for the `@for` loop which will be executed if the loop has no items to iterate over.
 - Code refactoring and naming improvements.
-
-------
 
 ## v1.0.0 (2024-02-26)